requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.6
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
filelock==3.14.0
identify==2.5.36
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
oauthlib==3.2.2
platformdirs==4.2.1
pre-commit==3.7.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
